requirements.txt
Flask
gunicorn
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-sdk
